#111720 — Mobile support for admin settings, forum description setting
Fixed the folder capitalization of skins and plugins as they were not previously recognized by the forum software.  Enabled the "show forum description" setting which allows for forum descriptions to be shown or hidden.  Styled forum description and mobile support for the admin settings page.  Replaced the default avatar thumbnail and side-picture with vectors.  Removed to-do field from the admin-only dashboard.  Added localization to various elements of the aforementioned page.  Release candidate.