Fix manifest.php
Previously the manifest wouldn't generate and it had an error about using $this in while not in object context. I just fixed this by determining whether or not the forum is using a custom icon and setting the newly created $icon variable accordingly. As far as I can determine the manifest works just fine now.