Last minute tweak
I had forgotten to include the skin config. The effect of this would be it always selecting Plastic as the set skin even when it wasn't. This is because the skin config variable was set in the default config and I forgot to import the config/skin.php which overrides that.